fine_art_description_generator/requirements.txt
streamlit==1.29.0
openai==1.6.1
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
numpy==1.24.3